feat: rename State to ComponentState
