+++ src/Docker/Configuration/ComponentState/Adapter.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Keboola\DockerBundle\Docker\Configuration\State;
+namespace Keboola\DockerBundle\Docker\Configuration\ComponentState;
 
 use Keboola\DockerBundle\Docker\Configuration;
 
 class Adapter extends Configuration\Adapter
 {
-    protected $configClass = Configuration\State::class;
+    protected $configClass = Configuration\ComponentState::class;
 
     protected function normalizeConfig($config)
     {
